+++ model/jpa/src/main/java/org/keycloak/models/jpa/entities/RealmEntity.java
@@ -33,8 +33,6 @@ import jakarta.persistence.NamedQueries;
 import jakarta.persistence.NamedQuery;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -152,7 +150,7 @@ public class RealmEntity {
     Collection<UserFederationMapperEntity> userFederationMappers = new LinkedList<>();
 
     @OneToMany(cascade ={CascadeType.PERSIST, CascadeType.REMOVE}, mappedBy = "realm")
-    List<OpenIdFederationEntity> openIdFederationList = new ArrayList<>();
+    List<OpenIdFederationEntity> openIdFederationList = List.of();
 
     @ElementCollection
     @MapKeyColumn(name="NAME")
@@ -608,7 +606,7 @@ public class RealmEntity {
     }
 
     public List<OpenIdFederationEntity> getOpenIdFederationList() {
-        return openIdFederationList == null ? new ArrayList<>() : openIdFederationList;
+        return openIdFederationList;
     }
 
     public void setOpenIdFederationList(List<OpenIdFederationEntity> openIdFederationList) {