+++ services/src/main/java/org/keycloak/utils/TrustChainProcessor.java
@@ -14,10 +14,9 @@ import java.util.Set;
 
 public interface TrustChainProcessor extends Provider {
 
-    List<TrustChainResolution> constructTrustChains(EntityStatement leafEs, Set<String> trustAnchorIds, boolean policyRequired, boolean forRp);
+    public TrustChainResolution constructTrustChains(EntityStatement leafEs, Set<String> trustAnchorIds, boolean forRp);
     <T extends EntityStatement> T parseAndValidateSelfSigned (String token, Class<T> clazz, JSONWebKeySet publicKey) throws IOException, JWSInputException, VerificationException;
     EntityStatement parseAndValidateSelfSigned(String token) throws InvalidTrustChainException;
     boolean validateEntityStatementFields(EntityStatement statement, String issuer, String subject);
-    TrustChainResolution findAcceptableMetadataPolicyChain(List<TrustChainResolution> trustChainResolutions, EntityStatement statement);
     JSONWebKeySet getKeySet();
 }